refactor: move failover component in api package
