+++ gravitee-apim-console-webui/src/management/api/failover/api-failover.component.html
