+++ gravitee-apim-console-webui/src/management/api/failover/api-failover.component.scss
@@ -2,7 +2,7 @@
 @use '@angular/material' as mat;
 @use '@gravitee/ui-particles-angular' as gio;
 
-@use '../../../../scss/gio-layout' as gio-layout;
+@use '../../../scss/gio-layout' as gio-layout;
 
 :host {
   @include gio-layout.gio-responsive-margin-container;
